load_function/requirements.txt
azure-functions
azure-storage-blob>=12.10.0

